Moved C# Project to CSharp Folder and Added F# Project
