+++ PlaybyPlayEFCore10/CSharp/myapi/WeatherEvent.cs
