+++ PlaybyPlayEFCore10/CSharp/myapi/appsettings.json
@@ -1,0 +1,14 @@
+{
+  "Logging": {
+    "IncludeScopes": false,
+    "LogLevel": {
+      "Default": "Debug",
+      "System": "Information",
+      "Microsoft": "Information"
+    }
+  },
+   "Data": { 
+       "PostgreConnection": { 
+          "ConnectionString": "User ID=julie;Password=12345;Host=localhost;Port=5432;Database=weatherdataRTM;Pooling=true;" 
+   }  }
+}
