+++ PlaybyPlayEFCore10/FSharp/README.md
@@ -1,0 +1,19 @@
+## EFCore 1.0 First Look
+
+This is the final project created by way of Modules 1-4 in the Pluralsight Play by Play: EF Core 1.0 First Look (https://app.pluralsight.com/library/courses/play-by-play-ef-core-1-0-first-look-julie-lerman/table-of-contents) released in Sept 2016.
+
+I will keep a "clean" version of the actual final state of the project from the course in a branch called PlayByPlayFinalCode.
+But I will likely evolve the example to include new learnings.
+
+1.0.1 was released just after we built this course so the first update (to master) will be to evolve this to 1.0.1. Since that was only bug fixes, the only thing that should change is package refs in the project.json file.
+
+The final modules (5 & 6) where I used an existing demo use code from a richer repo which is here: https://github.com/julielerman/EFCore-ASPNetCore-WebAPI-RTM
+
+
+
+## F# Version
+
+The F# sample was converted verbatim with minimal tweaks from the original. This version uses SQL Server instead of Postgres like the original code. I shouldn't be a problem as its virtually
+the same workflow.
+
+The Code First Migrations had to be converted by hand. Following this pattern, It shouldn't be hard to convert future migrations
