+++ PlaybyPlayEFCore10/FSharp/myapi/Dockerfile
@@ -1,0 +1,13 @@
+FROM microsoft/dotnet:latest
+
+COPY . /app
+
+WORKDIR /app
+
+RUN ["dotnet", "restore"]
+
+RUN ["dotnet", "build"]
+
+EXPOSE 5000/tcp
+
+CMD ["dotnet", "run", "--server.urls", "http://*:5000"]
