+++ PlaybyPlayEFCore10/FSharp/myapi/appsettings.json
@@ -1,0 +1,15 @@
+{
+  "Logging": {
+    "IncludeScopes": false,
+    "LogLevel": {
+      "Default": "Debug",
+      "System": "Information",
+      "Microsoft": "Information"
+    }
+  },
+  "Data":{
+    "SqlServerConnection":{
+      "ConnectionString":"Data Source=(local);Initial Catalog=WeatherDB;Persist Security Info=True;User ID=admin1;Password=admin1;MultipleActiveResultSets=True"
+    }
+  }
+}
